backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
duckdb==1.1.3
yfinance>=1.2.0
requests==2.32.3
apscheduler==3.10.4
python-dotenv==1.0.1
anthropic==0.40.0
transformers==4.47.1
torch==2.5.1
pandas==2.2.3
numpy==2.0.2
httpx==0.28.1
newspaper3k==0.2.8
lxml_html_clean>=0.1.0
feedparser==6.0.11
